Use static metadata for google_calendar
